libsemanage: silence -Wunused-variable warnings in SWIG wrapper
FIXME: A better way would be to override swig default?

In functions with an unsigned int* as their last parameter which is used
as a result counter, SWIG automatically creates a local variable, named
res$argnum (with $argnum being the index of the parameter, c.f.
https://github.com/swig/swig/blob/rel-3.0.10/Lib/typemaps/inoutlist.swg#L177)

This results in the following warnings reported by clang:

    semanageswig_wrap.c:6583:7: error: unused variable 'res3'
    [-Werror,-Wunused-variable]
      int res3 = SWIG_TMPOBJ ;
          ^
    semanageswig_wrap.c:6849:7: error: unused variable 'res3'
    [-Werror,-Wunused-variable]
      int res3 = SWIG_TMPOBJ ;
          ^
    semanageswig_wrap.c:7084:7: error: unused variable 'res3'
    [-Werror,-Wunused-variable]
      int res3 = SWIG_TMPOBJ ;
          ^
    ...

Silent these warnings by marking the variable used.