[TO TEST!] mcstrans: ensure mls_level_to_string uses memory consistently
mls_level_to_string() computes the length needed to craft a string and
fill it with calls to sprintf() before terminating with
"*(result + len) = 0;". If anything is different in this function,
compared to mls_compute_string_len(), this would lead to overwriting
structures in the heap and may lead to bugs which are hard to debug.

Instead of letting the heap be corrupted, make an assert fail if there
is an error in the logic of mls_level_to_string().

By the way, this solves a warning reported by clang's static analyzer
about assigning a value to "p" without using it, in the last "p += ...".

Signed-off-by: Nicolas Iooss <nicolas.iooss@m4x.org>