scripts: add a helper script to run Infer static analyzer
Facebook's Infer can perform static analysis on C code and generate
reports as text and HTML files. Using it on a project is as simple as
running "infer run -- make", but because SELinux's userspace is not so
straight to build, more commands are needed.

Add scripts/run-infer in order to run Infer on the project. This script
generates reports in scripts/infer-out/ (which is the default name of
Infer's output directory). As the reports contain many false positives
in categories "DEAD_STORE", "MEMORY_LEAK" and "RESOURCE_LEAK", mainly
because many constructors are not detected as such, the script disables
these categories (using --disable-issue-type ...).

Wire up this script with Circle-CI configuration. Doing so, each push
triggers a scan that generates artifacts that are browsable and hosted
by Circle-CI. Here is an example of such artifacts:
https://366-118970575-gh.circle-artifacts.com/0/infer-out/report.html/index.html

Signed-off-by: Nicolas Iooss <nicolas.iooss@m4x.org>