libsepol: replace sepol_av_to_string() with sepol_format_av()
sepol_av_to_string() is not thread-safe because it uses a static char
array. So no function using it is thread-safe, which is avoidable if a
buffer was given to sepol_av_to_string() instead (like "reentrant
functions" in glibc).

Introduce sepol_format_av() to format an access vector into a string.
Export it in libsepol.so so that it becomes standard API that does not
require linking with libsepol.a to be used.

In order to make users of sepol_av_to_string() migrate to this new
function, deprecate it.

Signed-off-by: Nicolas Iooss <nicolas.iooss@m4x.org>