python: make gettext setup compatible with Python 3.10 [TODO] TEST TRANSLATION [TODO] Migrate other gettext instantiation
According to Python's official documentation
https://docs.python.org/3.8/library/gettext.html#gettext.install,
the codeset parameter of gettext.install has been deprecated since
Python 3.8 and will be removed in Python 3.10.

This parameter was useful to support Python 2, but as the Python
libraries were migrated to only support Python 3, drop this
compatibility layer.

Moreover "import gettext" should always be available and it is unclear
when gettext.install might throw an exception, in Python 3. Therefore
remove the try/expect constructions around the setup of gettext, which
greatly simplifies the code.

Signed-off-by: Nicolas Iooss <nicolas.iooss@m4x.org>