libsepol/cil: do not output full blockinherit statements in secil2tree -A resolve
When running secil2tree -A resolve secilc/test/policy.cil, the output
contains:

    (block test_ba
        (blockinherit ba)
            (roletype test_ba.r test_ba.t)
            (blockabstract z.ba)
        )                   ; <-- This parenthesis is wrong
        (role test_ba.r)
        (type test_ba.t)
    )

The parentheses are not well balanced, because blockinherit has child
nodes even though this was not expected. Modify the CIL policy writer in
order to only produce a comment when a blockinherit was resolved. The
new output is:

    (block test_ba
        ( ; blockinherit ba
            (roletype test_ba.r test_ba.t)
            (blockabstract z.ba)
        )
        (role test_ba.r)
        (type test_ba.t)
    )

Signed-off-by: Nicolas Iooss <nicolas.iooss@m4x.org>